open buckets only after gatehring enough matches

+++ iris-mpc/src/server/mod.rs
@@ -538,7 +538,7 @@ pub async fn server_main(config: Config) -> eyre::Result<()> {
         request_parallelism: config.hawk_request_parallelism,
         connection_parallelism: config.hawk_connection_parallelism,
         disable_persistence: config.cpu_disable_persistence,
-        match_distances_buffer_size: config.match_distances_buffer_size as u32,
+        match_distances_buffer_size: config.match_distances_buffer_size,
         n_buckets: config.n_buckets,
     };
